Exercicio 2: DONE.
