+++ semana19/arquitetura/requests.rest
@@ -16,7 +16,7 @@ POST http://localhost:3003/login
 Content-Type: application/json
 
 {
-    "email": "nino@mail.com",
+    "email": "stark-millionaire@avengers.com",
     "password": "123456"
 }
