+++ semana19/arquitetura/src/data/loginQuery.ts
@@ -1,8 +1,7 @@
 import connection from "./connection";
-import { user } from "../model/user";
 
 // LOGIN
-export const login = async (email: string, password: string): Promise<any> => {
+export const login = async (email: string): Promise<any> => {
   const result = await connection("users")
     .select("*")
     .where("email", `${email}`);
